requirements.txt
fastapi==0.108.0
uvicorn==0.29.0
openai==1.24.0
neo4j==5.15.0
semantic-kernel==0.9.8b1
python-dotenv==1.0.1
tiktoken==0.5.2